Moved GraphQL Client integration into core Sentry package
