+++ Sentry.sln
@@ -174,10 +174,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Samples.GraphQL.Serv
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Samples.GraphQL.Client.Http", "samples\Sentry.Samples.GraphQL.Client.Http\Sentry.Samples.GraphQL.Client.Http.csproj", "{B01C5D8F-62EE-4E63-AE96-745BA1D2E175}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.GraphQL.Client", "src\Sentry.GraphQL.Client\Sentry.GraphQL.Client.csproj", "{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.GraphQL.Client.Tests", "test\Sentry.GraphQL.Client.Tests\Sentry.GraphQL.Client.Tests.csproj", "{1897129B-BA05-49C7-B37D-0328752C4A7C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -420,14 +416,6 @@ Global
 		{B01C5D8F-62EE-4E63-AE96-745BA1D2E175}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B01C5D8F-62EE-4E63-AE96-745BA1D2E175}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B01C5D8F-62EE-4E63-AE96-745BA1D2E175}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1897129B-BA05-49C7-B37D-0328752C4A7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1897129B-BA05-49C7-B37D-0328752C4A7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1897129B-BA05-49C7-B37D-0328752C4A7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1897129B-BA05-49C7-B37D-0328752C4A7C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -495,8 +483,6 @@ Global
 		{D62E79F4-FC3C-4D75-ABFE-CDE76EF46DDE} = {77454495-55EE-4B40-A089-71B9E8F82E89}
 		{B3BFB7BA-1A5E-468F-8C47-F0841AA75848} = {77454495-55EE-4B40-A089-71B9E8F82E89}
 		{B01C5D8F-62EE-4E63-AE96-745BA1D2E175} = {77454495-55EE-4B40-A089-71B9E8F82E89}
-		{A976354F-8451-4A6B-8BAC-CF94C2FCC6A7} = {AF6AF4C7-8AA2-4D59-8064-2D79560904EB}
-		{1897129B-BA05-49C7-B37D-0328752C4A7C} = {83263231-1A2A-4733-B759-EEFF14E8C5D5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0C652B1A-DF72-4EE5-A98B-194FE2C054F6}
