--- src/Sentry.GraphQL.Client/Usings.cs
@@ -1,5 +1,0 @@
-global using GraphQL;
-global using Sentry.Extensibility;
-global using Sentry.Internal;
-global using Sentry.Internal.Extensions;
-global using Sentry.Protocol;
