--- test/Sentry.GraphQL.Client.Tests/Sentry.GraphQL.Client.Tests.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-      <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
-    </PropertyGroup>
-  
-    <ItemGroup>
-      <ProjectReference Include="..\Sentry.Testing\Sentry.Testing.csproj" />
-    </ItemGroup>
-
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'net7.0'">
-    <!-- Use the net7.0 target to test netcoreapp3.1 -->
-    <ProjectReference Include="..\..\src\Sentry.GraphQL.Client\Sentry.GraphQL.Client.csproj" AdditionalProperties="TargetFramework=netcoreapp3.1" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-    <!-- Use the net6.0 target to test netstandard2.0 -->
-    <ProjectReference Include="..\..\src\Sentry.GraphQL.Client\Sentry.GraphQL.Client.csproj" AdditionalProperties="TargetFramework=netstandard2.0" />
-  </ItemGroup>
-
-</Project>
