--- test/Sentry.GraphQL.Client.Tests/Usings.cs
@@ -1,1 +1,0 @@
-global using Sentry.Testing;
