+++ test/Sentry.Tests/GraphQlRequestContentExtractorTests.cs
@@ -1,6 +1,4 @@
-using Sentry.Internal.Http;
-
-namespace Sentry.GraphQL.Client.Tests;
+namespace Sentry.Tests;
 
 public class GraphQlRequestContentExtractorTests
 {
