+++ test/Sentry.Tests/SentryGraphQlTestHelpers.cs
@@ -1,4 +1,4 @@
-namespace Sentry.GraphQL.Client.Tests;
+namespace Sentry.Tests;
 
 internal static class SentryGraphQlTestHelpers
 {
